Somehow I had turned of globals refactoring in Main, which breaks everything. Also, excluded sources should be relative to source dirs, so I changed that. If the regex has ^...$ it is assumed to be absolute.

+++ tests/Coriolis/refactored-src/src/coriolis.f95
@@ -21,7 +21,7 @@
         integer :: j
       real, parameter :: pi=4.*atan(1.)
       real, parameter :: freq=-2.*pi/(24.*3600.)
-      real, parameter :: f=2*freq
+      integer, parameter :: f=2*freq
       real, parameter :: dt=24.*3600./200.
       real, parameter :: alpha=f*dt
       real, parameter :: beta=0.25*alpha*alpha
